Add Exchange Event, Addressbook, Group, Person, EMail, Folder classes
