+++ app/logic/Calendar/ActiveSync/ActiveSyncEvent.ts
@@ -1,8 +1,8 @@
-import { Event } from "../Event";
 import { Participant } from "../Participant";
 import { InvitationResponse, type InvitationResponseInMessage } from "../Invitation/InvitationStatus";
 import { Frequency, Weekday, RecurrenceRule } from "../RecurrenceRule";
 import { IANAToWindowsTimezone, WindowsToIANATimezone } from "../ICal/WindowsTimezone";
+import { ExchangeEvent } from "../EWS/ExchangeEvent";
 import type { ActiveSyncCalendar } from "./ActiveSyncCalendar";
 import { ActiveSyncOutgoingInvitation } from "./ActiveSyncOutgoingInvitation";
 import { ActiveSyncError } from "../../Mail/ActiveSync/ActiveSyncError";
@@ -22,7 +22,7 @@ const ActiveSyncResponse: Record<InvitationResponseInMessage, number> = {
   [InvitationResponse.Decline]: 3,
 };
 
-export class ActiveSyncEvent extends Event {
+export class ActiveSyncEvent extends ExchangeEvent {
   declare calendar: ActiveSyncCalendar;
   declare parentEvent: ActiveSyncEvent;
   declare readonly exceptions: ArrayColl<ActiveSyncEvent>;
