+++ app/logic/Calendar/EWS/ExchangeEvent.ts
@@ -1,0 +1,4 @@
+import { Event } from "../Event";
+
+export class ExchangeEvent extends Event {
+}
