+++ app/logic/Contacts/EWS/EWSPerson.ts
@@ -1,5 +1,6 @@
-import { Person, ContactEntry } from '../../Abstract/Person';
+import { ContactEntry } from '../../Abstract/Person';
 import { StreetAddress } from '../StreetAddress';
+import { ExchangePerson } from "./ExchangePerson";
 import type { EWSAddressbook } from './EWSAddressbook';
 import { EWSCreateItemRequest } from "../../Mail/EWS/Request/EWSCreateItemRequest";
 import { EWSDeleteItemRequest } from "../../Mail/EWS/Request/EWSDeleteItemRequest";
@@ -7,7 +8,7 @@ import { EWSUpdateItemRequest } from "../../Mail/EWS/Request/EWSUpdateItemReques
 import { sanitize } from "../../../../lib/util/sanitizeDatatypes";
 import { ensureArray } from "../../util/util";
 
-export class EWSPerson extends Person {
+export class EWSPerson extends ExchangePerson {
   declare addressbook: EWSAddressbook | null;
 
   get itemID() {
