+++ app/logic/Contacts/EWS/ExchangeAddressbook.ts
@@ -1,0 +1,5 @@
+import { Addressbook } from "../Addressbook";
+
+export class ExchangeAddressbook extends Addressbook {
+  canSync: boolean = true;
+}
