+++ app/logic/Contacts/EWS/ExchangeGroup.ts
@@ -1,0 +1,4 @@
+import { Group } from '../../Abstract/Group';
+
+export class ExchangeGroup extends Group {
+}
