+++ app/logic/Contacts/EWS/ExchangePerson.ts
@@ -1,0 +1,5 @@
+import { Person } from '../../Abstract/Person';
+
+export class ExchangePerson extends Person {
+}
+
