+++ app/logic/Contacts/OWA/OWAAddressbook.ts
@@ -1,20 +1,20 @@
-import { Addressbook, type AddressbookShareCombinedPermissions } from "../Addressbook";
+import { type AddressbookShareCombinedPermissions } from "../Addressbook";
 import type { PersonUID } from "../../Abstract/PersonUID";
+import { ExchangeAddressbook } from "../EWS/ExchangeAddressbook";
 import { OWAPerson } from "./OWAPerson";
 import { OWAGroup } from "./OWAGroup";
 import { type OWAAccount, kMaxFetchCount } from "../../Mail/OWA/OWAAccount";
 import { owaGetPermissionsRequest, owaSetFolderPermissionsRequest } from "../../Mail/OWA/Request/OWAFolderRequests";
 import { owaFindPersonsRequest, owaGetPersonaRequest } from "./Request/OWAPersonRequests";
-import { getSharedPersons, ExchangePermission, deleteExchangePermissions, setExchangePermissions } from "../../Mail/EWS/EWSFolder";
+import { getSharedPersons, ExchangePermission, deleteExchangePermissions, setExchangePermissions } from "../../Mail/EWS/ExchangeFolder";
 import { RunOnce } from "../../util/flow/RunOnce";
 import { sanitize } from "../../../../lib/util/sanitizeDatatypes";
 import type { ArrayColl } from "svelte-collections";
 
-export class OWAAddressbook extends Addressbook {
+export class OWAAddressbook extends ExchangeAddressbook {
   readonly protocol: string = "addressbook-owa";
   /** Exchange FolderID for this addressbook. Not DistinguishedFolderId */
   folderID: string;
-  canSync: boolean = true;
   declare readonly persons: ArrayColl<OWAPerson>;
   declare readonly groups: ArrayColl<OWAGroup>;
   protected listContactsRunOnce = new RunOnce();
