+++ app/logic/Mail/OWA/Request/OWAFolderRequests.ts
@@ -1,7 +1,7 @@
 import { OWARequest } from "./OWARequest";
 import type { OWAEMail } from "../OWAEMail";
-import type { ExchangePermission } from "../../EWS/EWSFolder";
-// import { IconIndexPidTag } from "../../EWS/EWSFolder";
+import type { ExchangePermission } from "../../EWS/ExchangeFolder";
+// import { IconIndexPidTag } from "../../EWS/ExchangeFolder";
 
 export function owaFindMsgsInFolderRequest(folderID: string, maxFetchCount: number): OWARequest {
   return new OWARequest("FindItem", {
